requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.2
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
daphne==4.0.0
Django==4.2.3
django-environ==0.10.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.2
Faker==19.2.0
flake8==6.0.0
googlemaps==4.10.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
msgpack==1.0.5
openapi-codec==1.3.2
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyOpenSSL==23.2.0
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
service-identity==23.1.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.6
uuid==1.30
zope.interface==6.0
